+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Opacity semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- Color repository semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- Color mode semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Space semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Tag component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Link component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)